+++ Src/FluentAssertions/Formatting/EnumerableValueFormatter.cs
@@ -8,13 +8,19 @@ namespace FluentAssertions.Formatting
     public class EnumerableValueFormatter : IValueFormatter
     {
         /// <summary>
+        /// The number of items to include when formatting this object.
+        /// </summary>
+        /// <remarks>The default value is 32.</remarks>
+        protected virtual int MaxItems { get; } = 32;
+
+        /// <summary>
         /// Indicates whether the current <see cref="IValueFormatter"/> can handle the specified <paramref name="value"/>.
         /// </summary>
         /// <param name="value">The value for which to create a <see cref="System.String"/>.</param>
         /// <returns>
         /// <c>true</c> if the current <see cref="IValueFormatter"/> can handle the specified value; otherwise, <c>false</c>.
         /// </returns>
-        public bool CanHandle(object value)
+        public virtual bool CanHandle(object value)
         {
             return value is IEnumerable;
         }
@@ -28,11 +34,10 @@ namespace FluentAssertions.Formatting
             {
                 string postfix = "";
 
-                const int maxItems = 32;
-                if (enumerable.Count > maxItems)
+                if (enumerable.Count > MaxItems)
                 {
-                    postfix = $", …{enumerable.Count - maxItems} more…";
-                    enumerable = enumerable.Take(maxItems).ToArray();
+                    postfix = $", …{enumerable.Count - MaxItems} more…";
+                    enumerable = enumerable.Take(MaxItems).ToArray();
                 }
 
                 return "{" + string.Join(", ", enumerable.Select((item, index) => formatChild(index.ToString(), item))) + postfix + "}";